Making in 64 linux SD Change: heads of DB record/files
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSELEKTOR/trunk@687 7d9feeff-4400-0410-95ed-9cb8b92a518d

+++ gems2/mods/submod/ipm_gamma.cpp
@@ -88,7 +88,7 @@ double TProfil::pH_via_hydroxyl( double x[], double Factor, int j)
 }
 
 /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
-#define  a(j,i) (*(pmp->A+(i)+(j)*pmp->N))
+#define  a(j,i) ((double)(*(pmp->A+(i)+(j)*pmp->N)))
 /* Calculation of dual chemical potentials and concentrations
 * for species (indexed jb to je) in a k-th phase.
 * Input arrays X, XF, XFA,  input factors: Factor, MMC
